moved stuff
abstraction ; i love u <3
